feature refactor - copy_remote_storage_content added as separate executable
