+++ src/kfs_lib/includes/copy_remote_storage_content.h
@@ -1,0 +1,9 @@
+#ifndef KFS_COPY_REMOTE_STORAGE_CONTENT_H
+#define KFS_COPY_REMOTE_STORAGE_CONTENT_H
+
+#include "kfs_common.h"
+
+void copy_content(unsigned char type, char *path);
+int copy_remote_storage_content();
+
+#endif //KFS_COPY_REMOTE_STORAGE_CONTENT_H
