+++ src/kfs_lib/includes/kfs_init.h
@@ -5,6 +5,7 @@
 #include <glusterfs/api/glfs.h>
 
 #include "kfs_common.h"
+#include "copy_remote_storage_content.h"
 
 #pragma once
